requirements.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
gevent==23.9.1
greenlet==3.0.1
grpcio==1.59.0
grpcio-tools==1.59.0
gunicorn==21.2.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
platformdirs==3.11.0
protobuf==4.24.4
tomli==2.0.1
Werkzeug==3.0.1
yapf==0.40.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
